Major update.
ImageJ code is finished and is a full re-rewrite of previous code, now only requiring to be run as a macro (with script parameters) instead of with ImageJ's batch macro function. Just bug fixing now or slight modifications.
R code for Image Analysis is looking great, the statistics and plotting package will come soon(TM)