Fixed ROI selection after thresholding
ROIs were not be selected after completion of the thresholding macro. This is because restacking the image did not reselect the ROI. In addition, cropping the ROI also resulted in a new ROI polygon to fit the space. Thus, the overlay (visible but not present on the ROI) is added to the ROI and then selected. This resolves both issues. Now, the coloc function also has this as a failsafe, especially in case the previous images do not have a selected ROI.