Bugfixing and updates to processing
Fixed bugs involved with removing outliers
Fixed bug with identify_outliers that resulted in not working if multiple region/pairs did not match with ordered input (removed ordering for this call)
Created ability to save information about written plots to a text file.
Few tweaks to increase speed of processing.